Refactor ufs journal tests.
