+++ core/server/common/src/main/java/alluxio/master/journal/ufs/UfsJournalSnapshot.java
@@ -12,6 +12,7 @@
 package alluxio.master.journal.ufs;
 
 import alluxio.underfs.UfsStatus;
+import com.google.common.annotations.VisibleForTesting;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -24,7 +25,8 @@ import javax.annotation.concurrent.ThreadSafe;
  * This class provides a snapshot of everything in the journal and some utility methods.
  */
 @ThreadSafe
-final class UfsJournalSnapshot {
+@VisibleForTesting
+public final class UfsJournalSnapshot {
   /** The committed checkpoints. */
   private final List<UfsJournalFile> mCheckpoints;
   /** The journal edit logs including the incomplete log. */
@@ -124,7 +126,8 @@ final class UfsJournalSnapshot {
    *
    * @return the current log
    */
-  static UfsJournalFile getCurrentLog(UfsJournal journal) throws IOException {
+  @VisibleForTesting
+  public static UfsJournalFile getCurrentLog(UfsJournal journal) throws IOException {
     List<UfsJournalFile> logs = new ArrayList<>();
     UfsStatus[] statuses = journal.getUfs().listStatus(journal.getLogDir().toString());
     if (statuses != null) {
