+++ core/server/common/src/test/java/alluxio/master/journal/ufs/UfsJournalFileTest.java
@@ -9,12 +9,9 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.server.ft.journal.ufs;
+package alluxio.master.journal.ufs;
 
-import alluxio.testutils.BaseIntegrationTest;
 import alluxio.master.NoopMaster;
-import alluxio.master.journal.ufs.UfsJournal;
-import alluxio.master.journal.ufs.UfsJournalFile;
 import alluxio.util.URIUtils;
 
 import org.junit.Assert;
@@ -32,7 +29,7 @@ import java.util.Random;
 /**
  * Unit tests for {@link UfsJournalFile}.
  */
-public final class UfsJournalFileTest extends BaseIntegrationTest {
+public final class UfsJournalFileTest {
   @Rule
   public TemporaryFolder mFolder = new TemporaryFolder();
