+++ core/server/common/src/test/java/alluxio/master/journal/ufs/UfsJournalReaderTest.java
@@ -9,17 +9,11 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.server.ft.journal.ufs;
+package alluxio.master.journal.ufs;
 
-import alluxio.testutils.BaseIntegrationTest;
 import alluxio.ConfigurationTestUtils;
 import alluxio.master.NoopMaster;
 import alluxio.master.journal.JournalReader;
-import alluxio.master.journal.ufs.UfsJournal;
-import alluxio.master.journal.ufs.UfsJournalCheckpointWriter;
-import alluxio.master.journal.ufs.UfsJournalFile;
-import alluxio.master.journal.ufs.UfsJournalLogWriter;
-import alluxio.master.journal.ufs.UfsJournalReader;
 import alluxio.proto.journal.Journal;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.util.URIUtils;
@@ -37,7 +31,7 @@ import java.net.URI;
 /**
  * Unit tests for {@link UfsJournalReader}.
  */
-public final class UfsJournalReaderTest extends BaseIntegrationTest {
+public final class UfsJournalReaderTest {
   private static final long CHECKPOINT_SIZE = 10;
   @Rule
   public TemporaryFolder mFolder = new TemporaryFolder();
